Refactor "InternalServices" class to "Internals" namespace
