+++ src/CSnakes.Runtime/Python/PyDictionary.cs
@@ -1,7 +1,7 @@
 using CSnakes.Runtime.CPython;
 using System.Collections;
 using System.Diagnostics.CodeAnalysis;
-using static CSnakes.Runtime.Python.InternalServices;
+using CSnakes.Runtime.Python.Internals;
 
 namespace CSnakes.Runtime.Python;
