+++ src/CSnakes.Runtime/Python/PyList.cs
@@ -1,6 +1,6 @@
 using CSnakes.Runtime.CPython;
 using System.Collections;
-using static CSnakes.Runtime.Python.InternalServices;
+using CSnakes.Runtime.Python.Internals;
 
 namespace CSnakes.Runtime.Python;
