+++ src/CSnakes.Tests/PythonStaticGeneratorTests/FormatClassFromMethods.test_args.approved.txt
@@ -110,7 +110,7 @@ public static class TestClassExtensions
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject c_pyObject = PyObject.From(c)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, b_pyObject, c_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -124,7 +124,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.CallWithArgs([a_pyObject, b_pyObject, ..args ?? []]);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -139,7 +139,7 @@ public static class TestClassExtensions
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject c_pyObject = PyObject.From(c)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.CallWithKeywordArguments([a_pyObject, ..args ?? []], ["b", "c"], [b_pyObject, c_pyObject], null);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -153,7 +153,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.CallWithKeywordArguments([a_pyObject, b_pyObject], [], [], kwargs);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -168,7 +168,7 @@ public static class TestClassExtensions
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject c_pyObject = PyObject.From(c)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.CallWithKeywordArguments([a_pyObject, b_pyObject, ..args ?? []], ["c"], [c_pyObject], kwargs);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -183,7 +183,7 @@ public static class TestClassExtensions
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject c_pyObject = PyObject.From(c)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.CallWithKeywordArguments([a_pyObject, b_pyObject, ..args ?? []], ["c"], [c_pyObject], null);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64.Import(__result_pyObject);
                 return __return;
             }
         }
