+++ src/CSnakes.Tests/PythonStaticGeneratorTests/FormatClassFromMethods.test_basic.approved.txt
@@ -134,7 +134,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, b_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Double.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Double.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -148,7 +148,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, b_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -162,7 +162,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, b_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Double.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Double.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -176,7 +176,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, b_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Double.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Double.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -189,7 +189,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_list_of_ints;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.List<long, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64>.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.List<long, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64>.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -203,7 +203,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, b_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -217,7 +217,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, b_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.List<string, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String>.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.List<string, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String>.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -231,7 +231,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, b_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Dictionary<string, long, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64>.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Dictionary<string, long, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64>.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -244,7 +244,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_bytes;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.ByteArray.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.ByteArray.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -259,7 +259,7 @@ public static class TestClassExtensions
                 using PyObject start_pyObject = PyObject.From(start)!;
                 using PyObject end_pyObject = PyObject.From(end)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, start_pyObject, end_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Sequence<long, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64>.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Sequence<long, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64>.Import(__result_pyObject);
                 return __return;
             }
         }
