+++ src/CSnakes.Tests/PythonStaticGeneratorTests/FormatClassFromMethods.test_defaults.approved.txt
@@ -118,7 +118,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_default_str_arg;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -131,7 +131,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_default_int_arg;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -144,7 +144,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_default_float_arg;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Double.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Double.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -158,7 +158,7 @@ public static class TestClassExtensions
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject b_pyObject = PyObject.From(b)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject, b_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -171,7 +171,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_optional_int;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Boolean.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Boolean.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -184,7 +184,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_optional_str;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Boolean.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Boolean.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -197,7 +197,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_optional_list;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Boolean.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Boolean.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -210,7 +210,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_optional_any;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Boolean.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Boolean.Import(__result_pyObject);
                 return __return;
             }
         }
