+++ src/CSnakes.Tests/PythonStaticGeneratorTests/FormatClassFromMethods.test_dicts.approved.txt
@@ -98,7 +98,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_dict_str_int;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Dictionary<string, long, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64>.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Dictionary<string, long, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64>.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -111,7 +111,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_dict_str_list_int;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Dictionary<string, IReadOnlyList<long>, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.List<long, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64>>.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Dictionary<string, IReadOnlyList<long>, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.List<long, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64>>.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -124,7 +124,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_dict_str_dict_int;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Dictionary<string, IReadOnlyDictionary<string, long>, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Dictionary<string, long, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64>>.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Dictionary<string, IReadOnlyDictionary<string, long>, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Dictionary<string, long, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64>>.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -137,7 +137,7 @@ public static class TestClassExtensions
                 PyObject __underlyingPythonFunc = this.__func_test_mapping;
                 using PyObject a_pyObject = PyObject.From(a)!;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call(a_pyObject);
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Mapping<string, long, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String, global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Int64>.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Mapping<string, long, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String, global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Int64>.Import(__result_pyObject);
                 return __return;
             }
         }
