+++ src/CSnakes.Tests/PythonStaticGeneratorTests/FormatClassFromMethods.test_exceptions.approved.txt
@@ -87,7 +87,7 @@ public static class TestClassExtensions
                 logger.LogDebug("Invoking Python function: {FunctionName}", "test_raise_python_exception");
                 PyObject __underlyingPythonFunc = this.__func_test_raise_python_exception;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call();
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -99,7 +99,7 @@ public static class TestClassExtensions
                 logger.LogDebug("Invoking Python function: {FunctionName}", "test_nested_python_exception");
                 PyObject __underlyingPythonFunc = this.__func_test_nested_python_exception;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call();
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String.Import(__result_pyObject);
                 return __return;
             }
         }
