+++ src/CSnakes.Tests/PythonStaticGeneratorTests/FormatClassFromMethods.test_pybind11.approved.txt
@@ -82,7 +82,7 @@ public static class TestClassExtensions
                 logger.LogDebug("Invoking Python function: {FunctionName}", "test_pybind11_function");
                 PyObject __underlyingPythonFunc = this.__func_test_pybind11_function;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call();
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.Boolean.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.Boolean.Import(__result_pyObject);
                 return __return;
             }
         }
