+++ src/CSnakes.Tests/PythonStaticGeneratorTests/FormatClassFromMethods.test_sys.approved.txt
@@ -87,7 +87,7 @@ public static class TestClassExtensions
                 logger.LogDebug("Invoking Python function: {FunctionName}", "test_sys_executable");
                 PyObject __underlyingPythonFunc = this.__func_test_sys_executable;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call();
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String.Import(__result_pyObject);
                 return __return;
             }
         }
@@ -99,7 +99,7 @@ public static class TestClassExtensions
                 logger.LogDebug("Invoking Python function: {FunctionName}", "test_sys_path");
                 PyObject __underlyingPythonFunc = this.__func_test_sys_path;
                 using PyObject __result_pyObject = __underlyingPythonFunc.Call();
-                var __return = global::CSnakes.Runtime.Python.InternalServices.PyObjectImporters.String.Import(__result_pyObject);
+                var __return = global::CSnakes.Runtime.Python.Internals.PyObjectImporters.String.Import(__result_pyObject);
                 return __return;
             }
         }
